requirements.txt
# RANGEN Project Dependencies
# Install: pip install -r requirements.txt

# 🚀 Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0

# 🔐 Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# 🔄 Workflow & Agents
langgraph>=0.0.10
langchain-core>=0.1.0

# 🤖 MCP (Model Context Protocol)
mcp>=1.0.0
langgraph>=0.0.10
langchain-core>=0.1.0

# 🌐 Networking & Clients
httpx>=0.25.0
requests>=2.31.0
aiohttp>=3.9.0
nest-asyncio>=1.5.0

# 🖥️ UI / Frontend
streamlit>=1.30.0

# 🛠️ Utilities
numpy>=1.24.0
pandas>=2.0.0
psutil>=5.9.0
watchdog>=3.0.0

# 🧪 Testing
pytest>=7.0.0
pytest-asyncio>=0.23.0

# 🤖 ML/AI
scikit-learn>=1.3.0
